+++ crates/concurrency/Cargo.toml
@@ -6,22 +6,17 @@ license.workspace = true
 repository.workspace = true
 rust-version.workspace = true
 homepage.workspace = true
-description = "Concurrency management for RustFS - timeout, locking, backpressure, and I/O scheduling"
-keywords = ["rustfs", "concurrency", "timeout", "backpressure", "scheduling"]
+description = "Shared concurrency contract types for RustFS - workload admission, queue snapshots, worker pools, and policy types"
+keywords = ["rustfs", "concurrency", "admission", "backpressure", "workers"]
 categories = ["concurrency", "filesystem"]
 
 [dependencies]
 # Internal crates
 rustfs-io-core = { workspace = true }
-rustfs-io-metrics = { workspace = true }
 serde = { workspace = true }
 
 # Async runtime
-tokio = { workspace = true, features = ["sync", "time", "rt"] }
-tokio-util = { workspace = true }
-
-# Error handling
-thiserror = { workspace = true }
+tokio = { workspace = true, features = ["sync"] }
 
 # Logging
 tracing = { workspace = true }
@@ -29,18 +24,4 @@ tracing = { workspace = true }
 [dev-dependencies]
 insta = { workspace = true }
 serde_json = { workspace = true }
-tokio = { workspace = true, features = ["test-util","macros","rt-multi-thread"] }
-
-[features]
-default = ["timeout", "lock", "deadlock", "backpressure", "scheduler"]
-
-# Feature modules
-timeout = []
-lock = []
-deadlock = []
-backpressure = []
-scheduler = []
-
-[package.metadata.docs.rs]
-all-features = true
-rustdoc-args = ["--cfg", "docsrs"]
+tokio = { workspace = true, features = ["test-util", "macros", "rt-multi-thread"] }
